Rename logger module to sentry
